Implemented inheritance abstract class
